# gestock/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Controle de Estoque",
        "description": "API para controle de patrimônio e estoque",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/arquivos-imagem/{id}": {
            "get": {
                "tags": [
                    "Arquivos de Imagem"
                ],
                "summary": "Exibir registro",
                "description": "Retorna os dados de um arquivo de imagem específico",
                "operationId": "arquivos-imagem.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do arquivo de imagem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArquivoImagem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Arquivos de Imagem"
                ],
                "summary": "Atualizar registro",
                "description": "Atualiza os dados de um arquivo de imagem existente",
                "operationId": "arquivos-imagem.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArquivoImagem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArquivoImagem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Arquivos de Imagem"
                ],
                "summary": "Excluir registro",
                "description": "Remove um arquivo de imagem do sistema",
                "operationId": "arquivos-imagem.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Registro excluído"
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/condicoes-operacionais/{id}": {
            "get": {
                "tags": [
                    "Condições Operacionais"
                ],
                "summary": "Exibir registro",
                "description": "Retorna os dados de um condição operacional específico",
                "operationId": "condicoes-operacionais.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do condição operacional",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CondicaoOperacionalEquipamento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Condições Operacionais"
                ],
                "summary": "Atualizar registro",
                "description": "Atualiza os dados de um condição operacional existente",
                "operationId": "condicoes-operacionais.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CondicaoOperacionalEquipamento"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CondicaoOperacionalEquipamento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Condições Operacionais"
                ],
                "summary": "Excluir registro",
                "description": "Remove um condição operacional do sistema",
                "operationId": "condicoes-operacionais.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Registro excluído"
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/equipamentos-patrimoniados/{id}": {
            "get": {
                "tags": [
                    "Equipamentos Patrimoniados"
                ],
                "summary": "Exibir registro",
                "description": "Retorna os dados de um equipamento patrimoniado específico",
                "operationId": "equipamentos-patrimoniados.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do equipamento patrimoniado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipamentoPatrimoniado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Equipamentos Patrimoniados"
                ],
                "summary": "Atualizar registro",
                "description": "Atualiza os dados de um equipamento patrimoniado existente",
                "operationId": "equipamentos-patrimoniados.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EquipamentoPatrimoniado"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipamentoPatrimoniado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Equipamentos Patrimoniados"
                ],
                "summary": "Excluir registro",
                "description": "Remove um equipamento patrimoniado do sistema",
                "operationId": "equipamentos-patrimoniados.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Registro excluído"
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/espacos-armazenamento/{id}": {
            "get": {
                "tags": [
                    "Espaços de Armazenamento"
                ],
                "summary": "Exibir registro",
                "description": "Retorna os dados de um espaço de armazenamento específico",
                "operationId": "espacos-armazenamento.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do espaço de armazenamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EspacoArmazenamento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Espaços de Armazenamento"
                ],
                "summary": "Atualizar registro",
                "description": "Atualiza os dados de um espaço de armazenamento existente",
                "operationId": "espacos-armazenamento.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EspacoArmazenamento"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EspacoArmazenamento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Espaços de Armazenamento"
                ],
                "summary": "Excluir registro",
                "description": "Remove um espaço de armazenamento do sistema",
                "operationId": "espacos-armazenamento.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Registro excluído"
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/itens-estoque/{id}": {
            "get": {
                "tags": [
                    "Itens de Estoque"
                ],
                "summary": "Exibir registro",
                "description": "Retorna os dados de um item de estoque específico",
                "operationId": "itens-estoque.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do item de estoque",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemEstoque"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Itens de Estoque"
                ],
                "summary": "Atualizar registro",
                "description": "Atualiza os dados de um item de estoque existente",
                "operationId": "itens-estoque.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ItemEstoque"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemEstoque"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Itens de Estoque"
                ],
                "summary": "Excluir registro",
                "description": "Remove um item de estoque do sistema",
                "operationId": "itens-estoque.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Registro excluído"
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/marcas-equipamento/{id}": {
            "get": {
                "tags": [
                    "Marcas de Equipamento"
                ],
                "summary": "Exibir registro",
                "description": "Retorna os dados de um marca de equipamento específico",
                "operationId": "marcas-equipamento.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do marca de equipamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaEquipamento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Marcas de Equipamento"
                ],
                "summary": "Atualizar registro",
                "description": "Atualiza os dados de um marca de equipamento existente",
                "operationId": "marcas-equipamento.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarcaEquipamento"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MarcaEquipamento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Marcas de Equipamento"
                ],
                "summary": "Excluir registro",
                "description": "Remove um marca de equipamento do sistema",
                "operationId": "marcas-equipamento.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Registro excluído"
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EspacoArmazenamento": {
                "description": "Espaço de armazenamento",
                "properties": {
                    "id": {
                        "description": "ID do espaço",
                        "type": "integer"
                    },
                    "unidade_organizacional_id": {
                        "description": "ID da unidade organizacional",
                        "type": "integer"
                    },
                    "nome": {
                        "description": "Nome do espaço",
                        "type": "string"
                    },
                    "descricao": {
                        "description": "Descrição do espaço",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "MarcaEquipamento": {
                "description": "Marca de equipamento",
                "properties": {
                    "id": {
                        "description": "ID da marca",
                        "type": "integer"
                    },
                    "nome": {
                        "description": "Nome da marca",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CondicaoOperacionalEquipamento": {
                "description": "Condição operacional de equipamento",
                "properties": {
                    "id": {
                        "description": "ID da condição",
                        "type": "integer"
                    },
                    "nome": {
                        "description": "Nome da condição",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ArquivoImagem": {
                "description": "Arquivo de imagem",
                "properties": {
                    "id": {
                        "description": "ID do arquivo",
                        "type": "integer"
                    },
                    "nome_arquivo": {
                        "description": "Nome original do arquivo",
                        "type": "string"
                    },
                    "caminho": {
                        "description": "Caminho de armazenamento",
                        "type": "string"
                    },
                    "mime_type": {
                        "description": "Tipo MIME do arquivo",
                        "type": "string",
                        "nullable": true
                    },
                    "tamanho": {
                        "description": "Tamanho do arquivo em bytes",
                        "type": "integer",
                        "nullable": true
                    },
                    "hash": {
                        "description": "Hash SHA-256 do arquivo",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Data de upload",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação (Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}